% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sn_export.R
\name{mod_sn_export_app}
\alias{mod_sn_export_app}
\title{A minimal shiny app used for categorising streets}
\usage{
mod_sn_export_app(
  country,
  gisco_id,
  language = tidywikidatar::tw_get_language()
)
}
\arguments{
\item{language}{}
}
\description{
A minimal shiny app used for categorising streets
}
\examples{

if (interactive) {
  mod_sn_export_app(
    country = "IT",
    gisco_id = "IT_022205",
  )
}
}
