% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sn_import.R
\name{mod_sn_import_app}
\alias{mod_sn_import_app}
\title{A minimal shiny app used for categorising streets}
\usage{
mod_sn_import_app(language, connection = NULL, enable = TRUE)
}
\arguments{
\item{language}{}
}
\description{
A minimal shiny app used for categorising streets
}
\examples{

if (interactive) {
  mod_sn_import_app()
}
}
