% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sn_show_basic_municipality.R
\name{mod_sn_show_basic_municipality_app}
\alias{mod_sn_show_basic_municipality_app}
\title{A minimal shiny app used for showing summary statistics}
\usage{
mod_sn_show_basic_municipality_app(
  gisco_id,
  gisco_label = NULL,
  country = NULL,
  connection = NULL,
  language = tidywikidatar::tw_get_language()
)
}
\arguments{
\item{gisco_id}{Identifier of the city.}

\item{country}{Two letter country code}

\item{language}{Two letter language code.}
}
\description{
A minimal shiny app used for showing summary statistics
}
\examples{

if (interactive) {
  mod_sn_show_basic_municipality_app(
    gisco_id = "IT_022205",
    gisco_label = "Trento",
    country = "IT"
  )
}
}
