% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_sn_street_info.R
\name{mod_sn_street_info_app}
\alias{mod_sn_street_info_app}
\title{A minimal shiny app used for categorising streets}
\usage{
mod_sn_street_info_app(
  street_name,
  gisco_id,
  country,
  connection = NULL,
  language = tidywikidatar::tw_get_language()
)
}
\arguments{
\item{language}{}
}
\description{
A minimal shiny app used for categorising streets
}
\examples{

if (interactive) {
  mod_sn_street_info_app(
    street_name = "Belvedere San Francesco",
    gisco_id = "IT_022205",
    country = "IT"
  )
}
}
