% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name}
\alias{sn_clean_street_name}
\title{Clean street name in order to keep only the}
\usage{
sn_clean_street_name(street_name, country)
}
\arguments{
\item{street_name}{A character vector.}

\item{country}{Name of country.}
}
\value{

}
\description{
Clean street name in order to keep only the
}
\examples{

sn_clean_street_name(
  street_name = c(
    "Adele-Schreiber-Krieger-Straße",
    "Achillesstraße",
    "Adlerbrücke",
    "Akazienweg"
  ),
  country = "Germany"
)
}
