% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name_croatian}
\alias{sn_clean_street_name_croatian}
\title{Clean Croatian street name}
\usage{
sn_clean_street_name_croatian(street_name)
}
\arguments{
\item{street_name}{Name of street, after having removed street part.}
}
\value{
A character vector of length one. A name that can more easily be searched.
}
\description{
Clean Croatian street name
}
\examples{

sn_clean_street_name_croatian("Joze Kljakovića")
}
