% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name_greek}
\alias{sn_clean_street_name_greek}
\title{Clean Greek street name}
\usage{
sn_clean_street_name_greek(street_name)
}
\arguments{
\item{street_name}{Name of street, after having removed street part.}
}
\value{
A character vector of lenght one. A name that can more easily be searched.
}
\description{
Clean Greek street name
}
\examples{

sn_clean_street_name_greek("Αριστοτέλους")
}
