% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name_polish}
\alias{sn_clean_street_name_polish}
\title{Clean Polish street name}
\usage{
sn_clean_street_name_polish(street_name)
}
\arguments{
\item{street_name}{Name of street, after having removed street part (e.g. "Aleja")}
}
\value{
A character vector of length one. A name that can more easily be searched.
}
\description{
Clean Polish street name
}
\examples{

sn_clean_street_name_polish("Pawła Edmunda Strzeleckiego")
}
