% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name_romanian}
\alias{sn_clean_street_name_romanian}
\title{Clean Romanian street names}
\usage{
sn_clean_street_name_romanian(street_name)
}
\arguments{
\item{street_name}{Name of street, after having removed street part (e.g. "Aleea")}
}
\value{
A character vector of length one. A name that can more easily be searched.
}
\description{
Clean Romanian street names
}
\examples{

sn_clean_street_name_romanian("Aleea Egalității")
}
