% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_clean_streetname.R
\name{sn_clean_street_name_ukrainian}
\alias{sn_clean_street_name_ukrainian}
\title{Clean Ukrainian street name}
\usage{
sn_clean_street_name_ukrainian(street_name)
}
\arguments{
\item{street_name}{Name of street, after having removed street part.}
}
\value{
A character vector of length one. A name that can more easily be searched.
}
\description{
Clean Ukrainian street name
}
\examples{

sn_clean_street_name_ukrainian("Володимира Висоцького")
sn_clean_street_name_ukrainian("Кисловодська")
sn_clean_street_name_ukrainian("Кисловодський")
sn_clean_street_name_ukrainian("Миколи Терещенка")
}
