% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sn_connect_to_db}
\alias{sn_connect_to_db}
\title{Return a connection to be used for db}
\usage{
sn_connect_to_db(connection = NULL, type = "osm", country = NULL)
}
\arguments{
\item{connection}{Defaults to NULL. If NULL, uses local SQLite database. If given, must be a connection object (see example)}

\item{country}{Defaults to NULL.}
}
\value{
A connection object.
}
\description{
Return a connection to be used for db
}
\examples{
\donttest{
if (interactive()) {
  cache_connection <- DBI::dbConnect(
    RSQLite::SQLite(), # or e.g. odbc::odbc(),
    Driver =  ":memory:", # or e.g. "MariaDB",
    Host = "localhost",
    database = "example_db",
    UID = "example_user",
    PWD = "example_pwd"
  )
  sn_connect_to_db(cache_connection)
}
}

}
