% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sn_create_data_folder}
\alias{sn_create_data_folder}
\title{Creates the folder where `streetnamer` stores data.}
\usage{
sn_create_data_folder(ask = TRUE)
}
\arguments{
\item{ask}{Logical, defaults to TRUE. If FALSE, and cache folder does not exist, it just creates it without asking (useful for interactive sessions).}
}
\description{
Creates the folder where `streetnamer` stores data.
}
\examples{
\dontrun{
sn_create_data_folder()
}
}
