% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_create_map.R
\name{sn_create_map}
\alias{sn_create_map}
\title{Create thematic maps}
\usage{
sn_create_map(
  streets_sf = NULL,
  gisco_id = NULL,
  scope = "base",
  unnamed_streets = FALSE,
  connection = NULL
)
}
\arguments{
\item{streets_sf}{}

\item{gisco_id}{}

\item{scope}{Defaults to "base". Available values include "gender", "year_of_birth"}
}
\description{
Create thematic maps
}
