% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sn_disconnect_from_db}
\alias{sn_disconnect_from_db}
\title{Ensure that connection to cache is disconnected consistently}
\usage{
sn_disconnect_from_db(
  cache = NULL,
  db_connection = NULL,
  disconnect_db = TRUE,
  country = NULL
)
}
\arguments{
\item{cache}{Defaults to NULL. If given, it should be given either TRUE or FALSE. Typically set with `tw_enable_cache()` or `tw_disable_cache()`.}

\item{db_connection}{Defaults to NULL. If NULL, and caching is enabled, `tidywikidatar` will use a local sqlite database. A custom connection to other databases can be given (see vignette `caching` for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
Nothing, used for its side effects.
}
\description{
Ensure that connection to cache is disconnected consistently
}
\examples{

sn_disconnect_from_db()
}
