% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_download_osm.R
\name{sn_download_osm}
\alias{sn_download_osm}
\title{Download OSM data for whole countries from Geofabrik.}
\usage{
sn_download_osm(countries, overwrite = FALSE)
}
\arguments{
\item{countries}{One or more country names. For details on available country names see the dataset included in this package: `sn_osm_countries`}

\item{overwrite}{Logical, defaults to FALSE. If true, downloads new files even if already present.}
}
\value{
Used only for its side effects (downloads osm data).
}
\description{
N.B. Names do not always correspond to official name of countries and may include different geographic entities.
For a full list of available "countries" as made available by Geofabrik, see the internal dataset `sn_osm_countries`.
Be considered in downloading file.
}
\examples{

\dontrun{
sn_download_osm(countries = "Romania")
sn_download_osm(countries = c("chile", "colombia"))
}
}
