% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_export_checked.R
\name{sn_export}
\alias{sn_export}
\title{Exports data for a country or a municipality}
\usage{
sn_export(
  gisco_id = NULL,
  country = NULL,
  street_name = NULL,
  keep_only_latest = TRUE,
  only_checked = TRUE,
  remove_ignored = TRUE,
  additional_properties = c("P39", "P509", "P140", "P611", "P411", "P241", "P410", "P97",
    "P607", "P27", "P172"),
  include_image_credits = TRUE,
  unlist = FALSE,
  export_folder = "sn_data_export",
  export_format = NULL,
  write_file = FALSE,
  cache = TRUE,
  language = tidywikidatar::tw_get_language(),
  overwrite_cache = FALSE,
  connection = NULL,
  disconnect_db = FALSE
)
}
\arguments{
\item{gisco_id}{Defaults to NULL. If given, a valid code starting with a two letter country code. Either `gisco_id` or `country` must be given.}

\item{country}{Defaults to NULL. If given, must be a two-letter code. Returns data only for the given country.}

\item{unlist}{Defaults to FALSE. If TRUE, all data are unlisted in place, with values separated by a `;`.}

\item{export_folder}{Defaults to `sn_data_export`, created if not existing, overwrite files by default. If set to TRUE, unlist is also automatically set to TRUE.}

\item{export_format}{Defaults to NULL. Valid values include "csv", "geojson", "rds", "rds_sf"}
}
\description{
Exports data for a country or a municipality
}
