% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_export_checked.R
\name{sn_export_checked}
\alias{sn_export_checked}
\title{Title}
\usage{
sn_export_checked(
  gisco_id = NULL,
  country = NULL,
  source = "fixed_csv",
  additional_properties = c("P39", "P509", "P140", "P611", "P411", "P241", "P410", "P97",
    "P607", "P27", "P172"),
  include_image_credits = TRUE,
  fixed_folder = "sn_data_fixed",
  cache = NULL,
  language = tidywikidatar::tw_get_language(),
  overwrite_cache = FALSE,
  cache_connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{gisco_id}{Defaults to NULL. If given, a valid code starting with a two letter country code. Either `gisco_id` or `country` must be given.}

\item{country}{Defaults to NULL. If given, must be a two-letter code. Returns data only for the given country.}

\item{source}{Source for data. Valid values include "database" and "fixed_csv".}

\item{fixed_folder}{Base location of folder where csv files with manual fixes are stored. They can be located in subfolders.}
}
\description{
Title
}
