% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get.R
\name{sn_extract_streets}
\alias{sn_extract_streets}
\title{Extract shape files of roads from previously downloaded}
\usage{
sn_extract_streets(countries)
}
\arguments{
\item{countries}{The name of one or more geographic entities from files typically previously downloaded with `sn_download_osm()`}
}
\value{
Nothing, used for its side effects (extracts shapefiles from country-level zip files)
}
\description{
Extract shape files of roads from previously downloaded
}
\examples{
\dontrun{
sn_extract_streets(countries = "Romania")
}

}
