% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_filter.R
\name{sn_filter_people}
\alias{sn_filter_people}
\title{Filter search result and keep only people}
\usage{
sn_filter_people(search, language = "ro", limit = 10)
}
\arguments{
\item{search}{A data frame generated by `sn_search_wiki()`, or a search query. If a data frame is given, language and limits are ignore.}

\item{language}{Language to be used for the search.}

\item{limit}{Maximum numbers of responses to be given.}
}
\value{
A data frame with three columns, `id`, `label`, and `description`; all rows refer to a human being.
}
\description{
Filter search result and keep only people
}
\examples{

\dontrun{
sn_search_wiki("Antonio Vivaldi") \%>\%
  sn_filter_people()
}

}
