% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_details.R
\name{sn_get_details}
\alias{sn_get_details}
\title{Get a pre-defined set of details about}
\usage{
sn_get_details(
  id,
  collapse_lists = FALSE,
  language = tidywikidatar::tw_get_language(),
  connection = NULL,
  cache = TRUE
)
}
\arguments{
\item{id}{Wikidata id}

\item{collapse_lists}{Defaults to FALSE. If TRUE, collapses multiple values in a single string, with each value separated by ";". Useful for exports in tabular format.}

\item{cache}{Logical, defaults to TRUE.}
}
\value{

}
\description{
Get a pre-defined set of details about
}
