% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_details.R
\name{sn_get_details_by_lau}
\alias{sn_get_details_by_lau}
\title{Gets details for a single city}
\usage{
sn_get_details_by_lau(
  gisco_id,
  collapse_lists = FALSE,
  language = tidywikidatar::tw_get_language(),
  search_language = NULL,
  response_language = tidywikidatar::tw_get_language(),
  check_named_after_original = TRUE,
  check_named_after = TRUE,
  manual_check_columns = FALSE,
  connection = NULL,
  base_folder = "sn_data",
  streets_sf = NULL,
  cache = TRUE,
  overwrite_cache = FALSE,
  disconnect_db = TRUE
)
}
\arguments{
\item{gisco_id}{}

\item{collapse_lists}{}

\item{language}{}

\item{search_language}{}

\item{response_language}{}

\item{check_named_after_original}{}

\item{check_named_after}{}

\item{manual_check_columns}{}

\item{connection}{}

\item{base_folder}{}

\item{streets_sf}{}

\item{cache}{}

\item{overwrite_cache}{}

\item{disconnect_db}{}
}
\description{
Gets details for a single city
}
\examples{

gisco_id <- "DE_11000000"
}
