% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_gender_label.R
\name{sn_get_gender_label}
\alias{sn_get_gender_label}
\title{Get gender label based on a Wikidata Q identifier}
\usage{
sn_get_gender_label(
  named_after_id,
  language = tidywikidatar::tw_get_language(),
  cache_connection = NULL,
  cache = TRUE
)
}
\arguments{
\item{named_after_id}{Wikidata Q identifier.}

\item{language}{to be passed to `tidywikidatar` for caching. Label is always returned in English.}

\item{cache_connection}{Defaults to NULL, Passed to `tidywikidatar`.}

\item{cache}{Defaults to TRUE. Passed to `tidywikidatar`.}
}
\description{
Get gender label based on a Wikidata Q identifier
}
\examples{

if (interactive()) {
  sn_get_gender_label("Q7186")
}
}
