% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_lau_street_names.R
\name{sn_get_gisco_id}
\alias{sn_get_gisco_id}
\title{Get gisco id from name of municipality and country}
\usage{
sn_get_gisco_id(name, country)
}
\arguments{
\item{name}{Name of municipality}

\item{country}{Two-letter country code.}
}
\value{

}
\description{
Warning: there are many instances where there are more than one municipality with the same name in a country.
}
\examples{

sn_get_gisco_id(name = "Trento", country = "IT")
}
