% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_lau_street_names.R
\name{sn_get_lau_street_names}
\alias{sn_get_lau_street_names}
\title{Retrieve cached item}
\usage{
sn_get_lau_street_names(
  gisco_id,
  country,
  connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{connection}{Defaults to NULL. If NULL, and caching is enabled, `tidywikidatar` will use a local sqlite database. A custom connection to other databases can be given (see vignette `caching` for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection open.}

\item{id}{A characther vector, must start with Q, e.g. "Q180099" for the anthropologist Margaret Mead. Can also be a data frame of one row, typically generated with `tw_search()` or a combination of `tw_search()` and `tw_filter_first()`.}

\item{language}{Defaults to language set with `tw_set_language()`; if not set, "en". Use "all_available" to keep all languages. For available language values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}
}
\value{
If data present in cache, returns a data frame with cached data.
}
\description{
Retrieve cached item
}
\examples{


tw_set_cache_folder(path = tempdir())
tw_enable_cache()
tw_create_cache_folder(ask = FALSE)

df_from_api <- tw_get(id = "Q180099", language = "en")

df_from_cache <- tw_get_cached_item(
  id = "Q180099",
  language = "en"
)
}
