% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_street_name_wikidata_id.R
\name{sn_get_street_name_wikidata_id}
\alias{sn_get_street_name_wikidata_id}
\title{Retrieve combination of municipality/street name/wikidata id and category from database.}
\usage{
sn_get_street_name_wikidata_id(
  gisco_id,
  street_name,
  country,
  language = tidywikidatar::tw_get_language(),
  connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{gisco_id}{A characther vector of length one, must correpond to a gisco id. See `sn_lau_by_nuts` for available values.}

\item{country}{A two letter country code. See `sn_lau_by_nuts` for available values.}

\item{language}{Defaults to language set with `tw_set_language()`; if not set, "en". Use "all_available" to keep all languages. For available language values, see https://www.wikidata.org/wiki/Help:Wikimedia_language_codes/lists/all}

\item{connection}{Defaults to NULL. If NULL, and caching is enabled, `streetnamer` will use a local sqlite database. A custom connection to other databases can be given (see vignette `caching` for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection open.}
}
\value{
If data present in cache, returns a data frame with cached data.
}
\description{
Retrieve combination of municipality/street name/wikidata id and category from database.
}
\examples{


sn_set_data_folder(path = tempdir())
sn_write_street_name_wikidata_id(
  gisco_id = "IT_022205",
  street_name = "Belvedere San Francesco",
  wikidata_id = "Q676555",
  country = "IT", 
  person = TRUE,
  category = NA,
  session = as.character(floor(runif(1)*1e20)), 
  time = Sys.time(),
  checked = TRUE
)

sn_get_street_name_wikidata_id(
  gisco_id = "IT_022205",
  street_name = "Belvedere San Francesco",
  country = "IT"
)
}
