% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sn_lau_by_nuts}
\alias{sn_lau_by_nuts}
\title{A dataset with all European LAUs, with the correspondent NUTS regions}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 98931 rows and 14 columns.
}
\source{
\url{https://edjnet.github.io/lau_centres/}
}
\usage{
sn_lau_by_nuts
}
\description{
Based on 2020 LAUs and 2021 NUTS regions.
}
\details{
For details, see https://edjnet.github.io/lau_centres/
}
\keyword{datasets}
