% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{sn_set_data_folder}
\alias{sn_set_data_folder}
\alias{sn_get_data_folder}
\title{Set folder for db data}
\usage{
sn_set_data_folder(path = NULL)

sn_get_data_folder(path = NULL)
}
\arguments{
\item{path}{A path to a location used for caching data. If the folder does not exist, it will be created.}
}
\value{
The path to the caching folder, if previously set; the same path as given to the function; or the default, `sn_data` is none is given.
}
\description{
Consider using a folder out of your current project directory, e.g. `sn_set_data_folder("~/R/sn_data/")`: you will be able to use the same cache in different projects, and prevent cached files from being sync-ed if you use services such as Nextcloud or Dropbox.
}
\examples{
\donttest{
if (interactive()) {
  sn_set_data_folder(fs::path(fs::path_home_r(), "R", "sn_data"))
}
}
sn_get_data_folder()
}
