% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_standard_country.R
\name{sn_standard_country}
\alias{sn_standard_country}
\title{Turn a country name its a standard format, either country name, or two letter country code}
\usage{
sn_standard_country(country, type = "name")
}
\arguments{
\item{country}{A character string. Either full name of a country, or a two-letter country code.}

\item{type}{Defines output format. If "name", outputs full name. If "code", outputs two letter country code.}
}
\description{
Turn a country name its a standard format, either country name, or two letter country code
}
\examples{

sn_standard_country("DE", type = "name")

sn_standard_country("France", type = "code")

# if requested input the same as output, it is returned consitently
sn_standard_country("FR", type = "code")
}
