% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get.R
\name{sn_subset_streets}
\alias{sn_subset_streets}
\title{Keep only roads within a given boundary.}
\usage{
sn_subset_streets(boundary, streets)
}
\arguments{
\item{boundary}{An object typically created with `get_city_boundaries()`}

\item{country}{The name of the country. Requested to ensure correct identification of city.}
}
\value{
A data frame of the sf class including all roads insidet the given boundary
}
\description{
Keep only roads within a given boundary.
}
\examples{

subset_roads(city_boundary, roads)
}
