% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_search_wiki.R
\name{sn_tidy_wiki}
\alias{sn_tidy_wiki}
\title{Return (almost) all information from a Wikidata item in a tidy format}
\usage{
sn_tidy_wiki(id, cache = TRUE)
}
\arguments{
\item{id}{A characther vector, must start with Q, e.g. "Q254" for Wolfgang Amadeus Mozart}

\item{cache}{Logical, defaults to TRUE. If TRUE, it stores all retrieved data in a local sqlite database.}
}
\value{
A data.frame (a tibble) with two columns: property and value
}
\description{
Return (almost) all information from a Wikidata item in a tidy format
}
\examples{
\dontrun{
sn_extract_wiki("Q254")
}

}
