% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_write_to_db.R
\name{sn_write_lau_street_names}
\alias{sn_write_lau_street_names}
\title{Writes item to cache}
\usage{
sn_write_lau_street_names(
  df,
  type,
  country,
  overwrite = FALSE,
  connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{df}{A data frame with two columns: gisco_id and name.}

\item{type}{Type of data to be stored, e.g "osm".}

\item{country}{Two letter country code.}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, it first deletes all rows associated with the item(s) included in `item_df`. Useful if the original Wikidata object has been updated.}

\item{connection}{Defaults to NULL. If NULL, and caching is enabled, `tidywikidatar` will use a local sqlite database. A custom connection to other databases can be given (see vignette `caching` for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{
Nothing, used for its side effects.
}
\description{
Writes item to cache. Typically used internally, but exported to enable custom caching solutions.
}
\examples{

tw_set_cache_folder(path = fs::path(tempdir(), paste(sample(letters, 24), collapse = "")))
tw_create_cache_folder(ask = FALSE)
tw_disable_cache()

df_from_api <- tw_get(id = "Q180099", language = "en")

df_from_cache <- tw_get_cached_item(
  id = "Q180099",
  language = "en"
)

is.null(df_from_cache) # expect TRUE, as nothing has yet been stored in cache

tw_write_item_to_cache(
  item_df = df_from_api,
  language = "en"
)

df_from_cache <- tw_get_cached_item(
  id = "Q180099",
  language = "en"
)

is.null(df_from_cache) # expect a data frame, same as df_from_api
}
