% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_street_name_wikidata_id.R
\name{sn_write_street_name_wikidata_id}
\alias{sn_write_street_name_wikidata_id}
\title{Writes wikidata id for given street name}
\usage{
sn_write_street_name_wikidata_id(
  gisco_id,
  street_name,
  country,
  wikidata_id,
  person,
  gender,
  category,
  tag,
  checked,
  session,
  time,
  overwrite = FALSE,
  append = TRUE,
  connection = NULL,
  disconnect_db = TRUE
)
}
\arguments{
\item{country}{Two letter country code.}

\item{overwrite}{Logical, defaults to FALSE. If TRUE, it first deletes all rows associated with the item(s) included in `item_df`. Useful if the original Wikidata object has been updated.}

\item{connection}{Defaults to NULL. If NULL, and caching is enabled, `streetnamer` will use a local sqlite database. A custom connection to other databases can be given (see vignette `caching` for details).}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}

\item{df}{A data frame with two columns: gisco_id and name.}
}
\value{
Nothing, used for its side effects.
}
\description{
Writes wikidata id for given street name
}
\examples{

sn_set_data_folder(path = tempdir())

sn_write_street_name_wikidata_id(
  gisco_id = "IT_022205",
  country = "IT",
  street_name = "Belvedere San Francesco",
  person = TRUE,
  wikidata_id = "Q676555",
  gender = "male",
  category = "religion",
  tag = "",
  checked = TRUE,
  session = "testing",
  time = Sys.time(),
  append = TRUE,
  overwrite = FALSE,
  disconnect_db = TRUE
)
sn_get_street_name_wikidata_id(
  gisco_id = "IT_022205",
  street_name = "Belvedere San Francesco",
  country = "IT"
)
}
