% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_details.R
\name{sn_get_details_by_country}
\alias{sn_get_details_by_country}
\title{Get a pre-defined set of details about}
\usage{
sn_get_details_by_country(
  country,
  collapse_lists = FALSE,
  language = tidywikidatar::tw_get_language(),
  connection = NULL,
  base_folder = "sn_data",
  streets_sf = NULL,
  cache = TRUE,
  overwrite_cache = FALSE,
  disconnect_db = TRUE
)
}
\arguments{
\item{country}{A country, either a two letter country code, or country full name. See `sn_country_codes` for valid values.}

\item{collapse_lists}{Defaults to FALSE. If TRUE, collapses multiple values in a single string, with each value separated by ";". Useful for exports in tabular format.}

\item{streets_sf}{Defaults to NULL. If given, used to speed up processing.
Must be an sf object such as the ones output by `ll_osm_get_roads()`.
Passed to `ll_osm_get_lau_streets()`.}

\item{cache}{Logical, defaults to TRUE.}

\item{disconnect_db}{Defaults to TRUE. If FALSE, leaves the connection to cache open.}
}
\value{

}
\description{
Get a pre-defined set of details about
}
