% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_details.R
\name{sn_get_details_by_country}
\alias{sn_get_details_by_country}
\title{Get a pre-defined set of details about}
\usage{
sn_get_details_by_country(
  country,
  collapse_lists = FALSE,
  language = tidywikidatar::tw_get_language(),
  connection = NULL,
  base_folder = "sn_data",
  cache = TRUE
)
}
\arguments{
\item{country}{A country, either a two letter country code, or country full name. See `sn_country_codes` for valid values.}

\item{collapse_lists}{Defaults to FALSE. If TRUE, collapses multiple values in a single string, with each value separated by ";". Useful for exports in tabular format.}

\item{cache}{Logical, defaults to TRUE.}
}
\value{

}
\description{
Get a pre-defined set of details about
}
