% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_details.R
\name{sn_get_details_by_lau}
\alias{sn_get_details_by_lau}
\title{Gets details for a single city}
\usage{
sn_get_details_by_lau(
  gisco_id,
  additional_properties = c("P39", "P509", "P140", "P611", "P411", "P241", "P410", "P97",
    "P607", "P27", "P172"),
  export_format = NULL,
  collapse_lists = FALSE,
  language = tidywikidatar::tw_get_language(),
  search_language = NULL,
  response_language = tidywikidatar::tw_get_language(),
  check_named_after_original = TRUE,
  check_named_after = TRUE,
  checked_df = NULL,
  manual_check_columns = FALSE,
  connection = NULL,
  connection_search = NULL,
  base_folder = "sn_data",
  streets_sf = NULL,
  street_names_df = NULL,
  cache = TRUE,
  overwrite_cache = FALSE,
  disconnect_db = TRUE
)
}
\arguments{
\item{gisco_id}{A gisco identifier of a European LAU.}

\item{additional_properties}{Defaults to a vector of Wikidata properties, with common properties of interest for humans. Some basic properties related to birth, death, and occupation are included by default.}

\item{export_format}{Defaults to NULL. If given, it exports outputs in the given format. Available values include "csv" and "geojson".}

\item{search_language}{Defaults to NULL, guessed based on country.}

\item{response_language}{Defaults to `tidywikidatar::tw_get_language()`. Used
for defining the language in which label and description are returned.}

\item{check_named_after_original}{Defaults to TRUE. If TRUE, a search is
performed on the original street name. If the first result has values for
the property "named after", this takes precedence over other methods.}

\item{check_named_after}{Defaults to TRUE. If TRUE, a search is performed on
the "cleaned" name of the street. If the result is a street, road, square,
or similar, and this has values for the property "named after", this is
kept instead of the standard method.}

\item{checked_df}{Defaults to NULL. If given, a data frame with a `street_name` and
`wikidata_id` column. Takes precedence over searches.}

\item{streets_sf}{Defaults to NULL. If given, used to speed up processing.
Must be an sf object such as the ones output by `ll_osm_get_roads()`.
Passed to `ll_osm_get_lau_streets()`.}

\item{street_names_df}{Defaults to NULL. If TRUE, must be a data frame with
two columns, named "street_name" and "name_clean" respectively. If given, these
are passed directly to the search routine. Useful when name cleaning
provided by the package is not satisfying, e.g. in places such as some
Belgian cities where street names are given in more than one language.}
}
\description{
Gets details for a single city
}
\examples{

gisco_id <- "DE_11000000"
}
