% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_info_box.R
\name{sn_get_info_box}
\alias{sn_get_info_box}
\title{Get data from Wikidata and generate a box with key information}
\usage{
sn_get_info_box(wikidata_id, language = "en")
}
\arguments{
\item{wikidata_id}{A wikidata identifiers, must start with "Q".}

\item{language}{Defaults to "en". Must correspond to a two letter-code recognised by Wikidata.}
}
\value{
A Shiny taglist / HTML code.
}
\description{
Get data from Wikidata and generate a box with key information
}
\examples{

sn_get_info_box("Q676555")
}
