% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_get_city_streets.R
\name{sn_get_named_city_streets}
\alias{sn_get_named_city_streets}
\title{Gets roads within city boundaries, crops them, and keeps only those with names}
\usage{
sn_get_named_city_streets(gisco_id, cache_names = FALSE, cache_sf = FALSE)
}
\arguments{
\item{gisco_id}{Gisco id of the municipality}
}
\value{

}
\description{
This function downloads all roads for a given country, downloads boundaries for a given administrative unit, and removes roads with no names.
}
