% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sn_search_dedicated_to.R
\name{sn_search_dedicated_to}
\alias{sn_search_dedicated_to}
\title{Try to guess to whom a street is dedicated to}
\usage{
sn_search_dedicated_to(gisco_id, search_language = NULL)
}
\arguments{
\item{gisco_id}{}

\item{search_language}{Defaults to NULL, guessed based on country.}
}
\value{

}
\description{
Try to guess to whom a street is dedicated to
}
